test(#zimic): split interceptor method tests by file
