+++ packages/zimic/src/interceptor/http/HttpInterceptor/types/classes.ts
@@ -1,0 +1,4 @@
+import BrowserHttpInterceptor from '../BrowserHttpInterceptor';
+import NodeHttpInterceptor from '../NodeHttpInterceptor';
+
+export type HttpInterceptorClass = typeof BrowserHttpInterceptor | typeof NodeHttpInterceptor;
